@@ -1,4 +1,7 @@
 apply plugin: 'com.android.application'
+apply plugin: 'kotlin-android'
+apply plugin: 'kotlin-android-extensions'
+apply plugin: 'kotlin-kapt'
 
 // Create a variable called keystorePropertiesFile, and initialize it to your
 // keystore.properties file, in the rootProject folder.
@@ -19,43 +22,54 @@ android {
             storePassword keystoreProperties['storePassword']
         }
     }
-    compileSdkVersion 25
-    buildToolsVersion "25.0.3"
-    dataBinding.enabled = true
+    compileSdkVersion 28
+
     defaultConfig {
         applicationId 'com.soleren.pythonsamples'
         minSdkVersion 16
-        targetSdkVersion 26
+        targetSdkVersion 28
         versionName '3.3'
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         signingConfig signingConfigs.config
         versionCode 9
+
         vectorDrawables {
             useSupportLibrary = true
         }
     }
     buildTypes {
+
         release {
-            minifyEnabled false
+            minifyEnabled true
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
             signingConfig signingConfigs.config
         }
-    }
-    productFlavors {
+        debug {
+            minifyEnabled false
+        }
     }
 }
 
 dependencies {
-    compile fileTree(include: ['*.jar'], dir: 'libs')
-    androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
+    implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
+
+    def androidXVer = '1.0.0'
+    implementation "androidx.core:core-ktx:$androidXVer"
+    implementation "androidx.appcompat:appcompat:$androidXVer"
+    implementation "com.google.android.material:material:$androidXVer"
+    implementation "androidx.legacy:legacy-support-v4:$androidXVer"
+    implementation "androidx.recyclerview:recyclerview:$androidXVer"
+    implementation "androidx.cardview:cardview:$androidXVer"
+
+    def constraintVer = '1.1.3'
+    implementation "androidx.constraintlayout:constraintlayout:$constraintVer"
+
+    def playServicesVer = '17.0.0'
+    implementation "com.google.android.gms:play-services-ads:$playServicesVer"
+
+    kaptAndroidTest('androidx.test.espresso:espresso-core:3.1.0-beta02', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
-    compile 'com.android.support:appcompat-v7:25.3.1'
-    compile 'com.android.support.constraint:constraint-layout:1.0.2'
-    compile 'com.android.support:design:25.3.1'
-    compile 'com.android.support:support-v4:25.3.1'
-    compile 'com.android.support:recyclerview-v7:25.3.1'
-    compile 'com.android.support:cardview-v7:25.3.1'
-    compile 'com.google.android.gms:play-services-ads:10.2.4'
-    testCompile 'junit:junit:4.12'
+    testImplementation 'junit:junit:4.12'
 }
